Again

+++ lib/pdfkit/middleware.rb
@@ -15,7 +15,7 @@ class PDFKit
       set_request_to_render_as_pdf(env) if render_as_pdf?
       status, headers, response = @app.call(env)
 
-      if File.exists?(render_to) && rendering_pdf?
+      if File.exists?(render_to) && rendering_pdf? && headers['Content-Type'] =~ /text\/html|application\/xhtml\+xml/
         file = File.open(render_to, "rb")
         body = file.read
         file.close
@@ -46,8 +46,8 @@ class PDFKit
 
           headers['Content-Length'] = (body.respond_to?(:bytesize) ? body.bytesize : body.size).to_s
           headers['Content-Type']   = 'application/pdf'
+          [status, headers, response]
         end
-        [status, headers, response]
       end
 
     end
